Move hook/event callbacks to new ST-Events folder
