+++ src/ST-Events/Triggers.cs
@@ -1,0 +1,125 @@
+using System.Text.RegularExpressions;
+using CounterStrikeSharp.API.Core;
+using CounterStrikeSharp.API.Modules.Utils;
+using CounterStrikeSharp.API.Modules.Memory.DynamicFunctions;
+
+namespace SurfTimer;
+
+public partial class SurfTimer
+{
+    // Trigger start touch handler - CBaseTrigger_StartTouchFunc
+    internal HookResult OnTriggerStartTouch(DynamicHook handler)
+    {
+        CBaseTrigger trigger = handler.GetParam<CBaseTrigger>(0);
+        CBaseEntity entity = handler.GetParam<CBaseEntity>(1);
+        CCSPlayerController client = new CCSPlayerController(new CCSPlayerPawn(entity.Handle).Controller.Value!.Handle);
+        
+        if (client.IsBot || !client.IsValid)
+        {
+            return HookResult.Continue;
+        }
+
+        else 
+        {
+            // Implement Trigger Start Touch Here
+            Player player = playerList[client.UserId ?? 0];
+            #if DEBUG
+            player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc -> {trigger.DesignerName} -> {trigger.Entity!.Name}");
+            #endif
+
+            if (trigger.Entity!.Name != null)
+            {
+                // Map end zones -- hook into map_end
+                if (trigger.Entity.Name == "map_end")
+                {
+                    // MAP END ZONE
+                    if (player.Timer.IsRunning)
+                    {
+                        player.Timer.Stop();
+                        player.Stats.PB[0,0] = player.Timer.Ticks;
+                        player.Controller.PrintToChat($"{PluginPrefix} You finished the map in {player.HUD.FormatTime(player.Stats.PB[0,0])}!");
+                        // player.Timer.Reset();
+                    }
+
+                    #if DEBUG
+                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Red}Map Stop Zone");
+                    #endif
+                }
+
+                // Map start zones -- hook into map_start, (s)tage1_start
+                else if (trigger.Entity.Name.Contains("map_start") || 
+                        trigger.Entity.Name.Contains("s1_start") || 
+                        trigger.Entity.Name.Contains("stage1_start")) 
+                {
+                    player.Timer.Reset();
+
+                    #if DEBUG
+                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Green}Map Start Zone");
+                    // player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc -> KeyValues: {trigger.Entity.KeyValues3}");
+                    #endif
+                }
+
+                // Stage start zones -- hook into (s)tage#_start
+                else if (Regex.Match(trigger.Entity.Name, "^s([1-9][0-9]?|tage[1-9][0-9]?)_start$").Success)
+                {
+                    int stage = Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value) - 1;
+                    player.Timer.Stage = stage;
+
+                    #if DEBUG
+                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Yellow}Stage {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} Start Zone");
+                    #endif
+                }
+            }
+
+            return HookResult.Continue;
+        }
+    }
+
+    // Trigger end touch handler - CBaseTrigger_EndTouchFunc
+    internal HookResult OnTriggerEndTouch(DynamicHook handler)
+    {
+        CBaseTrigger trigger = handler.GetParam<CBaseTrigger>(0);
+        CBaseEntity entity = handler.GetParam<CBaseEntity>(1);
+        CCSPlayerController client = new CCSPlayerController(new CCSPlayerPawn(entity.Handle).Controller.Value!.Handle);
+        
+        if (client.IsBot || !client.IsValid)
+        {
+            return HookResult.Continue;
+        }
+
+        else
+        {
+            // Implement Trigger End Touch Here
+            Player player = playerList[client.UserId ?? 0];
+            #if DEBUG
+            player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_EndTouchFunc -> {trigger.DesignerName} -> {trigger.Entity!.Name}");
+            #endif
+
+            if (trigger.Entity!.Name != null)
+            {
+                // Map start zones -- hook into map_start, (s)tage1_start
+                if (trigger.Entity.Name.Contains("map_start") || 
+                    trigger.Entity.Name.Contains("s1_start") || 
+                    trigger.Entity.Name.Contains("stage1_start")) 
+                {
+                    // MAP START ZONE
+                    player.Timer.Start();
+
+                    #if DEBUG
+                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.LightRed}EndTouchFunc{ChatColors.Default} -> {ChatColors.Green}Map Start Zone");
+                    #endif
+                }
+
+                // Stage start zones -- hook into (s)tage#_start
+                else if (Regex.Match(trigger.Entity.Name, "^s([1-9][0-9]?|tage[1-9][0-9]?)_start$").Success)
+                {
+                    #if DEBUG
+                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.LightRed}EndTouchFunc{ChatColors.Default} -> {ChatColors.Yellow}Stage {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} Start Zone");
+                    #endif
+                }
+            }
+
+            return HookResult.Continue;
+        }
+    }
+}
